+++ internal/ui/src/i18n.js
@@ -1,7 +1,7 @@
 import i18n from 'i18next';
 import { initReactI18next } from 'react-i18next';
-import de from './locales/de.json' assert { type: 'json' };
-import en from './locales/en.json' assert { type: 'json' };
+import de from './locales/de.json';
+import en from './locales/en.json';
 
 i18n
   .use(initReactI18next)